Language.LTL

Description

A library for LTL formulae. Contains a parser and printer to spot's and spin's format and a generator for LTL formulae that can generate random forumlae of a given size and enumerate formulae of a given size.

Abstract Representation of LTL Formulae

data Formula

type for LTL formulae

abstract representation of ! (X ((F (a | b)) U (1 & (G c))) )

Unary Neg $
  Unary X $ 
    Binary U 
      (Unary F $ 
        Binary Or (AP "a") (AP "b") ) 
      (Binary And 
        (Nullary True) 
        (Unary G $ AP "c"))

Constructors

AP Text

Nullary Bool

Unary UnOp Formula

Binary BinOp Formula Formula

newtype SpinFormula

type for LTL formulae to be printed in spin's format

Constructors

SpinFormula Formula

Parser and Printer

parse :: Text -> Either String Formula

the parser of one LTL formula in spot's format

>>> parse "(p) U (G q)"
Right (Binary U (AP "p") (Unary G (AP "q")))

class ToText a where

printer of LTL formulae in spot's and spin's format

>>> toText $ (Binary U (AP "p") (Unary G (AP "q")))
"(p U (G q))"

>>> toText $ SpinFormula  (Binary U (AP "p") (Unary G (AP "q")))
"(p U ([] q))"

Methods

toText :: a -> Text

Properties of LTL Formulae

module Language.LTL.Property

LTL Generator

forms :: Select m => [BinOp] -> [UnOp] -> [Text] -> Gen m Formula

select a formula of a given size

the allowed binary and unary operators and the atomic propositions are given

>>> forms [And, Or, U] [F, G, X, Neg] ["p0","p1"] 3
Unary G (Unary F (AP "p1"))

forms_nf :: Select m => ([Text] -> Gen m Formula) -> Int -> Gen m Formula

select a formula of a given size where the atomic propositions are numbered increasingly

a generator for formulae and the maximal number of atomic propositions used are given (if the number of atomic propositions is at least as large as the desired size we obtain all formulae of this size when enumerating)

>>> forms_nf (forms [And,Or,U] [F,G,X,Neg]) 5 15
Unary Neg 
  (Binary U 
    (Unary Neg 
      (Binary And 
        (Binary U   (Unary X (AP "p0"))   (AP "p0")) 
        (Binary U 
          (Unary G (Unary X (AP "p1"))) 
          (Unary X (AP "p2"))
        )
      )			
    ) 
    (AP "p3")
  )

select :: Select m => [t] -> m t

Random Formulae

forms_uniform_nf :: Select m => Int -> Gen m Formula

select formulae where the operators are chosen uniformely

forms_uniform_nf = forms_nf (forms [minBound..maxBound][minBound..maxBound])

forms_fg_nf :: Select m => Int -> Gen m Formula

seleft formulae of the F,G fragment of LTL (only the temporal operators F and G are allowed)

forms_fg_nf = forms_nf (forms ([minBound..maxBound] L.\\ [U]) [F,G,Neg])

Generate a random formula of size 3 with at most 2 atomic propositions:

>>> (forms_fg_nf 2 3) :: IO Formula
Unary G (Unary F (AP "p0"))

Generate all formulae of size 3 with at most 2 atomic propositions:

>>> (forms_fg_nf 2 3) :: [Formula]
[Binary Or (AP "p0") (AP "p1")
,Binary Or (AP "p0") (AP "p0")
,Binary And (AP "p0") (AP "p1")
,Binary And (AP "p0") (AP "p0")
,Unary G (Unary F (AP "p0"))
,Unary Neg (Unary F (AP "p0"))
,Unary F (Unary G (AP "p0"))
,Unary Neg (Unary G (AP "p0"))
,Unary F (Unary Neg (AP "p0"))
,Unary G (Unary Neg (AP "p0"))
]

forms_morefg_nf :: Select m => Int -> Gen m Formula

select formulae where the operators F and G occur with higher probability

forms_morefg_nf = forms_nf (forms [minBound..maxBound] ([minBound..maxBound]++[F,G]))

Patterns

e :: Int -> Formula

u :: Int -> Formula

u2 :: Int -> Formula

r :: Int -> Formula

c1 :: Int -> Formula

c2 :: Int -> Formula

q :: Int -> Formula

s :: Int -> Formula

f :: Int -> Formula