Language.LTL.Gen

Description

A generator for LTL formulae. Random formulae of a given size can be computed and all formulae of this size can be enumerated. Some patterns are also implemented.

General Generators

class Monad m => Select m where

typeclass for different selections of a list: pick one randomly (in IO monad), or select all (in list monad)

an instance (selector) represents a set of elements

>>> select [Neg, F, G, X]) :: IO UnOp
F

>>> (select [Neg, F, G, X]) :: [UnOp]
[Neg,F,G,X]

Methods

select :: [t] -> m t

type Gen m a = Int -> m a

represents a collection of sets indexed by numbers

Here a set indexed by s contains elements of size s.

gen_unop :: Select m => Gen m [UnOp]

a generator for unary operators

>>> (gen_unop 2) :: IO [UnOp]
[G,F]

>>> (gen_unop 2) :: [[UnOp]]
[ [Neg,Neg],[Neg,F],[Neg,G],[Neg,X]
, [F,Neg],[F,F],[F,G],[F,X]
, [G,Neg],[G,F],[G,G],[G,X]
, [X,Neg],[X,F],[X,G],[X,X]
]

gen_binop :: Select m => Gen m [BinOp]

a generator for binary operators

times :: Select m => Gen m a -> Gen m b -> Gen m (a, b)

cross product of collections represented by generators where the sizes of the componentes are added

>>> times gen_unop gen_binop 3
([X],[U,And])

unions :: Select m => [m a] -> m a

unions of sets represented by selectors

>>> unions  [gen_unop 1, gen_unop 2] :: IO [UnOp ]
[G]

Generators for Formulae

Enumeration and Random Formulae

forms_unary :: Select m => [BinOp] -> [UnOp] -> [Text] -> Gen m Formula

select formula of given size beginning with an unary operator s.th. neither F F, G G or Neg Neg occur

the allowed binary and unary operators and the atomic propositions are given

forms :: Select m => [BinOp] -> [UnOp] -> [Text] -> Gen m Formula

select a formula of a given size

the allowed binary and unary operators and the atomic propositions are given

>>> forms [And, Or, U] [F, G, X, Neg] ["p0","p1"] 3
Unary G (Unary F (AP "p1"))

forms_nf :: Select m => ([Text] -> Gen m Formula) -> Int -> Gen m Formula

select a formula of a given size where the atomic propositions are numbered increasingly

a generator for formulae and the maximal number of atomic propositions used are given (if the number of atomic propositions is at least as large as the desired size we obtain all formulae of this size when enumerating)

>>> forms_nf (forms [And,Or,U] [F,G,X,Neg]) 5 15
Unary Neg 
  (Binary U 
    (Unary Neg 
      (Binary And 
        (Binary U   (Unary X (AP "p0"))   (AP "p0")) 
        (Binary U 
          (Unary G (Unary X (AP "p1"))) 
          (Unary X (AP "p2"))
        )
      )			
    ) 
    (AP "p3")
  )

aps_gen :: (Select m, Eq a) => Int -> [a] -> [a] -> m [a]

generate a given number of atomic propositions that are numbered increasingly z_i <= 1 + max { z1,..., z_(i-1) }

>>> aps_gen 3 [] [1, 2, 3, 4]
[1,2,1]

forms_uniform_nf :: Select m => Int -> Gen m Formula

select formulae where the operators are chosen uniformely

forms_uniform_nf = forms_nf (forms [minBound..maxBound][minBound..maxBound])

forms_fg_nf :: Select m => Int -> Gen m Formula

seleft formulae of the F,G fragment of LTL (only the temporal operators F and G are allowed)

forms_fg_nf = forms_nf (forms ([minBound..maxBound] L.\\ [U]) [F,G,Neg])

Generate a random formula of size 3 with at most 2 atomic propositions:

>>> (forms_fg_nf 2 3) :: IO Formula
Unary G (Unary F (AP "p0"))

Generate all formulae of size 3 with at most 2 atomic propositions:

>>> (forms_fg_nf 2 3) :: [Formula]
[Binary Or (AP "p0") (AP "p1")
,Binary Or (AP "p0") (AP "p0")
,Binary And (AP "p0") (AP "p1")
,Binary And (AP "p0") (AP "p0")
,Unary G (Unary F (AP "p0"))
,Unary Neg (Unary F (AP "p0"))
,Unary F (Unary G (AP "p0"))
,Unary Neg (Unary G (AP "p0"))
,Unary F (Unary Neg (AP "p0"))
,Unary G (Unary Neg (AP "p0"))
]

forms_morefg_nf :: Select m => Int -> Gen m Formula

select formulae where the operators F and G occur with higher probability

forms_morefg_nf = forms_nf (forms [minBound..maxBound] ([minBound..maxBound]++[F,G]))

Patterns

c1 :: Int -> Formula

c2 :: Int -> Formula

e :: Int -> Formula

q :: Int -> Formula

r :: Int -> Formula

s :: Int -> Formula

u :: Int -> Formula

u2 :: Int -> Formula

f :: Int -> Formula