Language.LTL.Ops

Description

Manipulating LTL formulae.

Documentation

children :: Formula -> [(Formula -> Formula, Formula)]

decompose formula into subformulae

pos :: Formula -> [Pos]

compute positions in the syntax tree of a formula

>>> pos $ Binary U (AP "p") (Unary G (AP "q"))
[[],[0],[1],[1,0]]

possubs :: Formula -> [(Pos, Formula)]

compute positions and subformulae at these positions

>>> possubs $ Binary U (AP "p") (Unary G (AP "q"))
[ ([],Binary U (AP "p") (Unary G (AP "q")))
, ([0],AP "p")
, ([1],Unary G (AP "q"))
, ([1,0],AP "q")
]

appos :: Formula -> [Pos]

postions where an atomic proposition is in a formula

>>> appos $ Binary U (AP "p") (Unary G (AP "q"))
[[0],[1,0]]

peek :: Formula -> Pos -> Formula

compute subformula at position

>>> peek (Binary U (AP "p") (Unary G (AP "q"))) [1]
Unary G (AP "q")

poke :: Formula -> (Pos, Formula) -> Formula

replace the subformula at the position with another formula

>>> poke (Binary U (AP "p") (Unary G (AP "q"))) ([1],AP "p")
Binary U (AP "p") (AP "p")

pokes :: Formula -> [(Pos, Formula)] -> Formula

replace multiple subformulae

pokes = foldl poke