Language.LTL.Parser

Description

A parser for LTL formulae in spot's format.

Documentation

parse :: Text -> Either String Formula

the parser of one LTL formula in spot's format

>>> parse "(p) U (G q)"
Right (Binary U (AP "p") (Unary G (AP "q")))

formula :: Parser Formula

expression parser for LTL formulae

atomic :: Parser Formula

parser for nullary operators, atomic propositions and formulae enclosed in parentheses

>>> parseOnly atomic "(p1)"
Right (AP "p1")

identifier :: Parser Text

parser for names of atomic propositions, removing withespace after it

>>> parseOnly identifier "p1   "
Right "p1"

embracedBy :: Parser a -> Text -> Text -> Parser a

parser for things enclosed in some strings

>>> parseOnly (embracedBy identifier "[" "]") "[ p1  ]"
Right "p1"

parens :: Parser a -> Parser a

parser for formulae enclosed in parentheses

parens p = embracedBy p "(" ")"

>>> parseOnly (parens atomic) "((p1))"
Right (AP "p1")

token :: Text -> Parser ()

parser for tokens, removing all withespace after it

whitespace :: Parser ()